+++ packages/riverpod/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Improved error message for missing `dependencies` (thanks to @ValentinVignal)
 - Fixed various typos in the documentation (thanks to @ValentinVignal)
+- Fix not disposing autoDispose family providers
 
 ## 2.3.5 - 2023-04-18