+++ packages/riverpod/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased major
+
+- Bumped minimum Dart SDK to >= 3.0.0-dev
+
 ## 2.3.6 - 2023-04-24
 
 - Improved error message for missing `dependencies` (thanks to @ValentinVignal)